rename type.ts => types.internal.ts
