+++ packages/react-async-states/src/context.ts
@@ -1,5 +1,5 @@
 import * as React from "react";
-import {AsyncStateContextValue} from "./types";
+import {AsyncStateContextValue} from "./types.internal";
 
 export const AsyncStateContext =
   React.createContext<AsyncStateContextValue | null>(null);
