+++ packages/react-async-states/src/types.internal.ts
