+++ packages/shared/index.ts
@@ -3,7 +3,7 @@ import {
   ProducerProps,
   ProducerSavedProps
 } from "../async-state";
-import {PartialUseAsyncStateConfiguration} from "react-async-states/src/types";
+import {PartialUseAsyncStateConfiguration} from "react-async-states/src/types.internal";
 
 export const __DEV__ = process.env.NODE_ENV !== "production";
